port: Implement gPTP-capable TLV signaling message processing
Implementation of the transmission and reception of gPTP-capable TLV
signaling messages is crucial for determining per-domain asCapable
when Common Mean Link Delay Service (CMLDS) is enabled, as mandated
by the IEEE802.1AS-2020 standard.

For transmission, each active PTP domain  now periodically sends
gPTP-capable signaling messages with the gPTP-capable TLV at
intervals specified by the `logGptpCapableMessageInterval`
configuration parameter (as per IEEE802.1AS-2020 sections 10.4.1.3
and 10.6.4.4).

For reception, the mechanism is responsible for setting per domain
neighborGptpCapable to TRUE  when a gPTP-capable TLV signaling
message specific to the corresponding PTP domain is received.
Conversely, it sets neighborGptpCapable to FALSE if such a message
is not received after gPtpCapableReceiptTimeout
(as per IEEE802.1AS-2020 sections 10.4.2.2 and 10.7.3.3, 14.8.41).

Signed-off-by: Tan Tee Min <tee.min.tan@linux.intel.com>
Signed-off-by: Ganesh Venkatesan <ganesh.venkatesan@intel.com>
Signed-off-by: Chwee-Lin Choong <chwee.lin.choong@intel.com>