port: Implement asCapableAcrossDomains and neighborGptpCapable
This commit implements asCapableAcrossDomains and neighborGptpCapable,
as specified in IEEE 802.1AS-2020 Section 11.2.2.

In scenario with multiple PTP domains, a domain is asCapable only
when both asCapableAcrossDomains and neighborGptpCapable are TRUE.
The determination of asCapableAcrossDomains relies on the
Common Mean Link Delay Service (CMLDS). Simultaneously,
neighborGptpCapable is set to TRUE upon receiving periodic
gPTP-Capable TLV signaling messages from the peer,
adhering to the logGptpCapableMessageInterval value set in the
message.

Signed-off-by: Tan Tee Min <tee.min.tan@linux.intel.com>
Signed-off-by: Chwee-Lin Choong <chwee.lin.choong@intel.com>